+++ results/verified/20260720T232223Z_DD-003_2ea8dad5_ae62f6c1f1/reproduce.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+set -eu
+exec make dd003-source-graphs